Separate the 'db_driver' module into parts
