+++ mm2src/common/indexed_db/drivers/object_store.rs
@@ -1,0 +1,272 @@
+use super::{construct_event_closure, DbTransactionError, DbTransactionResult, ItemId};
+use crate::mm_error::prelude::*;
+use crate::stringify_js_error;
+use futures::channel::mpsc;
+use futures::StreamExt;
+use serde_json::Value as Json;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::Arc;
+use wasm_bindgen::prelude::*;
+use wasm_bindgen::JsCast;
+use web_sys::{IdbObjectStore, IdbRequest};
+
+pub struct IdbObjectStoreImpl {
+    pub(crate) object_store: IdbObjectStore,
+    pub(crate) aborted: Arc<AtomicBool>,
+}
+
+impl !Send for IdbObjectStoreImpl {}
+
+impl IdbObjectStoreImpl {
+    pub fn aborted(&self) -> bool { self.aborted.load(Ordering::Relaxed) }
+
+    pub async fn add_item(&self, item: &Json) -> DbTransactionResult<ItemId> {
+        if self.aborted.load(Ordering::Relaxed) {
+            return MmError::err(DbTransactionError::TransactionAborted);
+        }
+
+        // The [`InternalItem::item`] is a flatten field, so if we add the item without the [`InternalItem::_item_id`] id,
+        // it will be calculated automatically.
+        let js_value = match JsValue::from_serde(item) {
+            Ok(value) => value,
+            Err(e) => return MmError::err(DbTransactionError::ErrorSerializingItem(e.to_string())),
+        };
+        let add_request = match self.object_store.add(&js_value) {
+            Ok(request) => request,
+            Err(e) => return MmError::err(DbTransactionError::ErrorUploadingItem(stringify_js_error(&e))),
+        };
+
+        if let Err(_error_event) = Self::wait_for_request_complete(&add_request).await {
+            self.aborted.store(true, Ordering::Relaxed);
+            let error = Self::error_from_failed_request(&add_request);
+            return MmError::err(DbTransactionError::ErrorUploadingItem(error));
+        }
+
+        Self::item_id_from_completed_request(&add_request)
+    }
+
+    pub async fn get_items(&self, index_str: &str, index_value: Json) -> DbTransactionResult<Vec<(ItemId, Json)>> {
+        if self.aborted.load(Ordering::Relaxed) {
+            return MmError::err(DbTransactionError::TransactionAborted);
+        }
+
+        let index = index_str.to_owned();
+        let index_value_js =
+            JsValue::from_serde(&index_value).map_to_mm(|e| DbTransactionError::ErrorSerializingIndex {
+                index: index.clone(),
+                description: e.to_string(),
+            })?;
+
+        let db_index = match self.object_store.index(index_str) {
+            Ok(index) => index,
+            Err(_) => return MmError::err(DbTransactionError::NoSuchIndex { index }),
+        };
+        let get_request = match db_index.get_all_with_key(&index_value_js) {
+            Ok(request) => request,
+            Err(e) => {
+                return MmError::err(DbTransactionError::InvalidIndex {
+                    index,
+                    index_value,
+                    description: stringify_js_error(&e),
+                })
+            },
+        };
+
+        if let Err(_error_event) = Self::wait_for_request_complete(&get_request).await {
+            self.aborted.store(true, Ordering::Relaxed);
+            let error = Self::error_from_failed_request(&get_request);
+            return MmError::err(DbTransactionError::ErrorGettingItems(error));
+        }
+
+        Self::items_from_completed_request(&get_request)
+    }
+
+    pub async fn get_item_ids(&self, index_str: &str, index_value: Json) -> DbTransactionResult<Vec<ItemId>> {
+        if self.aborted.load(Ordering::Relaxed) {
+            return MmError::err(DbTransactionError::TransactionAborted);
+        }
+
+        let index = index_str.to_owned();
+        let index_value_js =
+            JsValue::from_serde(&index_value).map_to_mm(|e| DbTransactionError::ErrorSerializingIndex {
+                index: index.clone(),
+                description: e.to_string(),
+            })?;
+
+        let db_index = match self.object_store.index(index_str) {
+            Ok(index) => index,
+            Err(_) => return MmError::err(DbTransactionError::NoSuchIndex { index }),
+        };
+        let get_request = match db_index.get_all_keys_with_key(&index_value_js) {
+            Ok(request) => request,
+            Err(e) => {
+                return MmError::err(DbTransactionError::InvalidIndex {
+                    index,
+                    index_value,
+                    description: stringify_js_error(&e),
+                })
+            },
+        };
+
+        if let Err(_error_event) = Self::wait_for_request_complete(&get_request).await {
+            self.aborted.store(true, Ordering::Relaxed);
+            let error = Self::error_from_failed_request(&get_request);
+            return MmError::err(DbTransactionError::ErrorGettingItems(error));
+        }
+
+        Self::item_ids_from_completed_request(&get_request)
+    }
+
+    pub async fn get_all_items(&self) -> DbTransactionResult<Vec<(ItemId, Json)>> {
+        if self.aborted.load(Ordering::Relaxed) {
+            return MmError::err(DbTransactionError::TransactionAborted);
+        }
+
+        let get_request = match self.object_store.get_all() {
+            Ok(request) => request,
+            Err(e) => return MmError::err(DbTransactionError::UnexpectedState(stringify_js_error(&e))),
+        };
+
+        if let Err(_error_event) = Self::wait_for_request_complete(&get_request).await {
+            self.aborted.store(true, Ordering::Relaxed);
+            let error = Self::error_from_failed_request(&get_request);
+            return MmError::err(DbTransactionError::ErrorGettingItems(error));
+        }
+
+        Self::items_from_completed_request(&get_request)
+    }
+
+    pub async fn replace_item(&self, _item_id: ItemId, item: Json) -> DbTransactionResult<ItemId> {
+        if self.aborted.load(Ordering::Relaxed) {
+            return MmError::err(DbTransactionError::TransactionAborted);
+        }
+
+        let item_with_key = InternalItem { _item_id, item };
+        let js_value = match JsValue::from_serde(&item_with_key) {
+            Ok(value) => value,
+            Err(e) => return MmError::err(DbTransactionError::ErrorSerializingItem(e.to_string())),
+        };
+        let replace_request = match self.object_store.put(&js_value) {
+            Ok(request) => request,
+            Err(e) => return MmError::err(DbTransactionError::ErrorUploadingItem(stringify_js_error(&e))),
+        };
+
+        if let Err(_error_event) = Self::wait_for_request_complete(&replace_request).await {
+            self.aborted.store(true, Ordering::Relaxed);
+            let error = Self::error_from_failed_request(&replace_request);
+            return MmError::err(DbTransactionError::ErrorUploadingItem(error));
+        }
+
+        Self::item_id_from_completed_request(&replace_request)
+    }
+
+    pub async fn delete_item(&self, item_id: ItemId) -> DbTransactionResult<()> {
+        if self.aborted.load(Ordering::Relaxed) {
+            return MmError::err(DbTransactionError::TransactionAborted);
+        }
+
+        let item_id = JsValue::from(item_id);
+
+        let delete_request = match self.object_store.delete(&item_id) {
+            Ok(request) => request,
+            Err(e) => return MmError::err(DbTransactionError::ErrorDeletingItems(stringify_js_error(&e))),
+        };
+
+        if let Err(_error_event) = Self::wait_for_request_complete(&delete_request).await {
+            self.aborted.store(true, Ordering::Relaxed);
+            let error = Self::error_from_failed_request(&delete_request);
+            return MmError::err(DbTransactionError::ErrorDeletingItems(error));
+        }
+
+        Ok(())
+    }
+
+    pub async fn clear(&self) -> DbTransactionResult<()> {
+        if self.aborted.load(Ordering::Relaxed) {
+            return MmError::err(DbTransactionError::TransactionAborted);
+        }
+
+        let clear_request = match self.object_store.clear() {
+            Ok(request) => request,
+            Err(e) => return MmError::err(DbTransactionError::ErrorDeletingItems(stringify_js_error(&e))),
+        };
+
+        if let Err(_error_event) = Self::wait_for_request_complete(&clear_request).await {
+            self.aborted.store(true, Ordering::Relaxed);
+            let error = Self::error_from_failed_request(&clear_request);
+            return MmError::err(DbTransactionError::ErrorDeletingItems(error));
+        }
+
+        Ok(())
+    }
+
+    async fn wait_for_request_complete(request: &IdbRequest) -> Result<JsValue, JsValue> {
+        let (tx, mut rx) = mpsc::channel(2);
+
+        let onsuccess_closure = construct_event_closure(Ok, tx.clone());
+        let onerror_closure = construct_event_closure(Err, tx.clone());
+
+        request.set_onsuccess(Some(onsuccess_closure.as_ref().unchecked_ref()));
+        request.set_onerror(Some(onerror_closure.as_ref().unchecked_ref()));
+
+        rx.next().await.expect("The request event channel must not be closed")
+    }
+
+    fn item_id_from_completed_request(request: &IdbRequest) -> DbTransactionResult<ItemId> {
+        let result_js_value = match request.result() {
+            Ok(res) => res,
+            Err(e) => return MmError::err(DbTransactionError::UnexpectedState(stringify_js_error(&e))),
+        };
+
+        result_js_value
+            .into_serde()
+            .map_to_mm(|e| DbTransactionError::ErrorDeserializingItem(e.to_string()))
+    }
+
+    fn items_from_completed_request(request: &IdbRequest) -> DbTransactionResult<Vec<(ItemId, Json)>> {
+        let result_js_value = match request.result() {
+            Ok(res) => res,
+            Err(e) => return MmError::err(DbTransactionError::UnexpectedState(stringify_js_error(&e))),
+        };
+
+        if result_js_value.is_null() || result_js_value.is_undefined() {
+            return Ok(Vec::new());
+        }
+
+        let items: Vec<InternalItem> = match result_js_value.into_serde() {
+            Ok(items) => items,
+            Err(e) => return MmError::err(DbTransactionError::ErrorDeserializingItem(e.to_string())),
+        };
+
+        Ok(items.into_iter().map(|item| (item._item_id, item.item)).collect())
+    }
+
+    fn item_ids_from_completed_request(request: &IdbRequest) -> DbTransactionResult<Vec<ItemId>> {
+        let result_js_value = match request.result() {
+            Ok(res) => res,
+            Err(e) => return MmError::err(DbTransactionError::UnexpectedState(stringify_js_error(&e))),
+        };
+
+        if result_js_value.is_null() || result_js_value.is_undefined() {
+            return Ok(Vec::new());
+        }
+
+        result_js_value
+            .into_serde()
+            .map_to_mm(|e| DbTransactionError::ErrorDeserializingItem(e.to_string()))
+    }
+
+    fn error_from_failed_request(request: &IdbRequest) -> String {
+        match request.error() {
+            Ok(Some(exception)) => format!("{:?}", exception),
+            _ => "Unknown".to_owned(),
+        }
+    }
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+struct InternalItem {
+    _item_id: ItemId,
+    #[serde(flatten)]
+    item: Json,
+}
